requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
idna==2.10
pyTelegramBotAPI==3.7.4
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
